Pause Dialog
Created a pause dialog to appear in the playScreen. THough once the pause dialog is created by the esc key the playscreen will be stuck at the point in which the pause key was hit. This wil continue after going to the main menu as well. Adding some labels to instructions as weel and inserted a ui skin for the dialog in the assets folder.