Fixed Dialog box issue and Created functional pausing
+ Fixed issue when trying to disable the dialog box. It needs to be
removed from the Stage, then it has to allow the input to be focused
onto the play screen, and unpause the game.

=======================
Pausing
=======================
+ Created a variable in play screen that sets the isPaused variable to
true, which disables the stepping for the world in the render() method.
The dialog box then pops up. Duh :P

+ I've created it in such a way that we can reuse the StartScreen and
PlayScreen, which does waste RAM, but saves alittle CPU time, when the
player is constantly pausing and going to the main menu.

+ This also allows the player to continue where they left off. In order
for them to restart the game, they would have to restart the
application. There is a simple fix to this (if we have time). Create a
restart button in the StartScreen that will get rid of the Player, and
create a new PlayScreen and assign that object to the play screen static
variable in the IOU class.

+ Don't care about recreating the Instruction and Next Instruction
Screens and QuitScreen at the moment.

+ Fixed an issue where another dialog box would pop up when user clicks
outside of window. Clicking outside calls the PlayScreen's pause().
Clicking back inside the window, calls the resume(). You can actually
test this by commenting out the RemoveDialog() in the pause() and
resume(). Launch game. In the play screen, press ESC. Move the dialog
box somewhere. Click outside the window. Bomb! Two dialogs.

===============
BUGS
===============

+ Just realized there is a minor issue when reusing the screens. Each
time we set the screen to X screen, it calls the show() method. This
means that it recreates the labels and stuff for the screens that have
the label creation in it. This defeats the purpose of saving the object
in IOU.

+ Do we care about Instruction and NextInstruction Screens? If not, then
we don't care about it in the IOU class.

+ Also found another bug when Swithcing through the Instruction menus
(which actually could be because the labels are drawing over each
other). When going back and forth, sometimes the previous button would
jump straight to the Start Screen, and sometimes the previous button
would work and the 'go to Main Screen' wouldn't. As a temporary fix, I
put everything in the show() inside the end of the constructor (since, I
figured the show() would occur after the constructor). This way, it
wouldn't overlap the labels.... we'll see.

+ Still need to fix when we resize the window, buttons are all messed up
(aka, not working). Maybe use percentages instead of hard code to
prevent scaling issues (to make the labels responsive and relative to
the size of the window)?