Corrected Physics World
+ Able to see Bodies now in Debug mode.
- To see this, comment out everything in-between the game.batch.begin()
and game.batch.end() in the render() method of the PlayScreen class.

+ Player starts at (-3.0m, +1.5m), so camera is at (0,0).
+ Jumping and moving and shooting feels better.
- Although the bullet's graphics are harsh on the eyes, because its
moving too fast.
- Due to frame rate? Or Rendering slowly?

+ Player checks if its within the boundary (for the user's safety).
+ I also realized that the player is relatively big for its size, but it
works.

==============
Note
==============
- I've realized that the bodys' origins (0,0 coordinates) starts at the
center of it's self.
- This playlist helped a lot. (YT videos by dermetfan)
-
https://www.youtube.com/watch?v=zDEQycnhaAc&list=PLXY8okVWvwZ2Ph9LKWiNBZ1GRAc_TyDru

-

==============
BUG
==============
+ Because I scaled everything to that of the world, rather than the
pixels, the pause dialog isn't scaling and positioning correctly
(probably because it's camera is not viewing the same as the game's
camera).
To see this issue:
- Run Play Screen
- Press RIGHT CTRL to zoom camera out (Yes, everything is suppose to be
small.)
- Jump by pressing [Space] -> Quickly press [ESC]
- You should see dialog box on top right area.
- Click anywhere (it will send to Main menu)

- I believe the dialog is centered in the normal camera view port
(1000x800) and is not being scaled properly.
- When you zoom to 1000x1000, you can see that the dialog box's right
border is exactly at the edge of the screen.
- I just don't know how to or what to scale Up or down to.
- the stage that is passed into the constructor is the same as the
PlayScreen
- I left a bunch of stuff in the commented code of PausePopUP's
constructor to hopefully get an idea of how to fix this issue.

+ Another issue with spawning multiple dialog boxes
- possibly because of scaling issues, if you pause, then click anywhere
it will send you to the Main menu.
- OR when implementing Function to press ESC to resume the game (if so,
could remove)
- FIX about issue first before dealing with this

================
DEBUG Shortcuts
================
[CAMERA]:
- ZOOM:
Scroll mouse up or down. Press LEFT CTRL and scroll mouse to scroll 10*
times faster
Press RIGHT CTRL: instally zoom out to a viewport of 1300x1300 (width,
height)
- this is to see the bug with the dialog box

- MOVEMENT:
move camera by 0.1 meters
[NUMPAD]
8,2: Up and Down
4,6: Left and Right
5  : Reset to original position
9  : Move to player's body's position

[PLAYER]:
(To be deleted)- maybe
- ROTATION:
Apply some rotating force to the player for demonstration.
Press Up or Down arrow keys