Added End Game scholarship calculations
Enabled the gameOverScreen to calculate the scholarship based on the total grade of the player. This amount is then subtracted from the final debt amount.
Add sprites to the instructions screen for the enter and spacebar as well as the progressbar sprite for nextInstructionScreen.
The main Menu will exit the app. Press 'Q' to get to the gameOverScreen.